requirements.txt
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-mgmt-network==25.0.0
google-cloud-asset==3.24.0
google-cloud-storage==2.13.0
google-cloud-compute==1.15.0
pyyaml==6.0.1
jinja2==3.1.2
click==8.1.7
colorama==0.4.6
tabulate==0.9.0